Widen the hicolormap realloc size to size_t before the multiply (CodeQL)
CodeQL flagged the two high-colormap REALLOC size expressions in
pgtypes/regex/regc_color.c (vendored PostgreSQL/Henry-Spencer regex) as
'multiplication may overflow int before conversion to unsigned long'. The
int*int product is in fact bounded by the INT_MAX/(...) guards just above each
site, so it cannot overflow, but the arithmetic is computed in int before the
* sizeof(color) widening. Cast the first operand to size_t so the whole size
expression is evaluated in size_t -- same value, no int intermediate. Marked
/* MEOS */ per the vendored-fix convention so it survives an upstream re-sync
and can be reported upstream.

+++ pgtypes/regex/regc_color.c
@@ -435,7 +435,7 @@ newhicolorrow(struct colormap *cm,
 			return 0;
 		}
 		newarray = (color *) REALLOC(cm->hicolormap,
-									 cm->maxarrayrows * 2 *
+									 (size_t) cm->maxarrayrows * 2 * /* MEOS: widen to size_t before the int multiply (overflow already guarded above) */
 									 cm->hiarraycols * sizeof(color));
 		if (newarray == NULL)
 		{
@@ -478,7 +478,7 @@ newhicolorcols(struct colormap *cm)
 		return;
 	}
 	newarray = (color *) REALLOC(cm->hicolormap,
-								 cm->maxarrayrows *
+								 (size_t) cm->maxarrayrows * /* MEOS: widen to size_t before the int multiply (overflow already guarded above) */
 								 cm->hiarraycols * 2 * sizeof(color));
 	if (newarray == NULL)
 	{
